move webhook processing to helper
